Добавлен диалог на очистку репозитория

+++ app/src/main/java/ru/barinov/notes/ui/dataManagerFragment/DataManagerViewModel.kt
@@ -1,6 +1,8 @@
 package ru.barinov.notes.ui.dataManagerFragment
 
 
+import android.os.Bundle
+import androidx.fragment.app.DialogFragment
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.liveData
@@ -9,6 +11,7 @@ import com.google.firebase.ktx.Firebase
 import ru.barinov.notes.domain.CloudRepository
 import ru.barinov.notes.domain.curentDataBase.NotesRepository
 import ru.barinov.notes.domain.room.DataBase
+import ru.barinov.notes.ui.AgreementDialogFragment
 
 
 class DataManagerViewModel(
@@ -23,17 +26,29 @@ class DataManagerViewModel(
     private val cloudDataBase: CloudRepository = cloudDataBase
     private val auth = auth
 
-    private val _liveData = MutableLiveData<Boolean>()
-    override val ld: LiveData<Boolean> = _liveData
+    private val _onRepositoryDeletion = MutableLiveData<DialogFragment>()
+    override val onRepositoryDeletion: LiveData<DialogFragment> = _onRepositoryDeletion
+
+    private val _repositoryIsCleanedMessage = MutableLiveData<Boolean>()
+    override val repositoryIsCleanedMessage: LiveData<Boolean> = _repositoryIsCleanedMessage
+
 
     override fun deleteAllNotes() {
-        repository.deleteAll()
-        Thread {
-            localDB.clearAllTables()
-        }.start()
-        if (auth.currentUser != null) {
-            //ToDo}
+        val confirmation = AgreementDialogFragment()
+        _onRepositoryDeletion.postValue(confirmation)
+    }
+
+    fun onRepoDeletion(result: Bundle, key: String) {
+
+        if (result.getBoolean(key)) {
+            repository.deleteAll()
+            Thread {
+                localDB.clearAllTables()
+            }.start()
+            if (auth.currentUser != null) {
+                //ToDo}
+            }
+            _repositoryIsCleanedMessage.postValue(true)
         }
-        _liveData.postValue(true)
     }
 }
